\name{bionetdata-package}
\alias{bionetdata-package}
\alias{bionetdata}
\docType{package}
\title{
bionetdata: Biological and chemical data networks
}
\description{
Data Package that includes several examples of chemical and biological data networks represented through adjacency matrices of a graph
}
\details{
\tabular{ll}{
Package: \tab bionetdata\cr
Type: \tab Package\cr
Version: \tab 1.0.1\cr
Date: \tab 2014-09-07\cr
License: \tab GPL (>= 2) \cr
LazyLoad: \tab yes\cr
}

}
\author{
\emph{Giorgio Valentini} and \emph{Matteo Re}

DI, Dipartimento di Scienze dell'Informazione

Universita' degli Studi di Milano

\email{\{valentini,re\}@di.unimi.it}

Maintainer: 
\emph{Matteo Re} 
}


\keyword{ package }

